Port TemplateParser to Dispatcher project
