+++ src/Microsoft.AspNetCore.Routing/Resources.resx
@@ -141,56 +141,11 @@
   <data name="DefaultInlineConstraintResolver_TypeNotConstraint" xml:space="preserve">
     <value>The constraint type '{0}' which is mapped to constraint key '{1}' must implement the '{2}' interface.</value>
   </data>
-  <data name="TemplateRoute_CannotHaveCatchAllInMultiSegment" xml:space="preserve">
-    <value>A path segment that contains more than one section, such as a literal section or a parameter, cannot contain a catch-all parameter.</value>
-  </data>
-  <data name="TemplateRoute_CannotHaveDefaultValueSpecifiedInlineAndExplicitly" xml:space="preserve">
-    <value>The route parameter '{0}' has both an inline default value and an explicit default value specified. A route parameter cannot contain an inline default value when a default value is specified explicitly. Consider removing one of them.</value>
-  </data>
-  <data name="TemplateRoute_CannotHaveConsecutiveParameters" xml:space="preserve">
-    <value>A path segment cannot contain two consecutive parameters. They must be separated by a '/' or by a literal string.</value>
-  </data>
-  <data name="TemplateRoute_CannotHaveConsecutiveSeparators" xml:space="preserve">
-    <value>The route template separator character '/' cannot appear consecutively. It must be separated by either a parameter or a literal value.</value>
-  </data>
-  <data name="TemplateRoute_CatchAllCannotBeOptional" xml:space="preserve">
-    <value>A catch-all parameter cannot be marked optional.</value>
-  </data>
-  <data name="TemplateRoute_OptionalCannotHaveDefaultValue" xml:space="preserve">
-    <value>An optional parameter cannot have default value.</value>
-  </data>
-  <data name="TemplateRoute_CatchAllMustBeLast" xml:space="preserve">
-    <value>A catch-all parameter can only appear as the last segment of the route template.</value>
-  </data>
-  <data name="TemplateRoute_InvalidLiteral" xml:space="preserve">
-    <value>The literal section '{0}' is invalid. Literal sections cannot contain the '?' character.</value>
-  </data>
-  <data name="TemplateRoute_InvalidParameterName" xml:space="preserve">
-    <value>The route parameter name '{0}' is invalid. Route parameter names must be non-empty and cannot contain these characters: '{{', '}}', '/'. The '?' character marks a parameter as optional, and can occur only at the end of the parameter. The '*' character marks a parameter as catch-all, and can occur only at the start of the parameter.</value>
-  </data>
-  <data name="TemplateRoute_InvalidRouteTemplate" xml:space="preserve">
-    <value>The route template cannot start with a '/' or '~' character.</value>
-  </data>
-  <data name="TemplateRoute_MismatchedParameter" xml:space="preserve">
-    <value>There is an incomplete parameter in the route template. Check that each '{' character has a matching '}' character.</value>
-  </data>
-  <data name="TemplateRoute_RepeatedParameter" xml:space="preserve">
-    <value>The route parameter name '{0}' appears more than one time in the route template.</value>
-  </data>
   <data name="RouteConstraintBuilder_ValidationMustBeStringOrCustomConstraint" xml:space="preserve">
     <value>The constraint entry '{0}' - '{1}' on the route '{2}' must have a string value or be of a type which implements '{3}'.</value>
   </data>
   <data name="RouteConstraintBuilder_CouldNotResolveConstraint" xml:space="preserve">
     <value>The constraint entry '{0}' - '{1}' on the route '{2}' could not be resolved by the constraint resolver of type '{3}'.</value>
-  </data>
-  <data name="TemplateRoute_UnescapedBrace" xml:space="preserve">
-    <value>In a route parameter, '{' and '}' must be escaped with '{{' and '}}'.</value>
-  </data>
-  <data name="TemplateRoute_OptionalParameterCanbBePrecededByPeriod" xml:space="preserve">
-    <value>In the segment '{0}', the optional parameter '{1}' is preceded by an invalid segment '{2}'. Only a period (.) can precede an optional parameter.</value>
-  </data>
-  <data name="TemplateRoute_OptionalParameterHasTobeTheLast" xml:space="preserve">
-    <value>An optional parameter must be at the end of the segment. In the segment '{0}', optional parameter '{1}' is followed by '{2}'.</value>
   </data>
   <data name="AttributeRoute_DifferentLinkGenerationEntries_SameName" xml:space="preserve">
     <value>Two or more routes named '{0}' have different templates.</value>
@@ -198,10 +153,13 @@
   <data name="UnableToFindServices" xml:space="preserve">
     <value>Unable to find the required services. Please add all the required services by calling '{0}.{1}' inside the call to '{2}' in the application startup code.</value>
   </data>
-  <data name="TemplateRoute_Exception" xml:space="preserve">
-    <value>An error occurred while creating the route with name '{0}' and template '{1}'.</value>
-  </data>
   <data name="DispatcherFeatureIsRequired" xml:space="preserve">
     <value>The '{0}' has no '{1}'. '{2}' requires a dispatcher.</value>
+  </data>
+  <data name="TemplateRoute_CannotHaveDefaultValueSpecifiedInlineAndExplicitly" xml:space="preserve">
+    <value>The route parameter '{0}' has both an inline default value and an explicit default value specified. A route parameter cannot contain an inline default value when a default value is specified explicitly. Consider removing one of them.</value>
+  </data>
+  <data name="TemplateRoute_Exception" xml:space="preserve">
+    <value>An error occurred while creating the route with name '{0}' and template '{1}'.</value>
   </data>
 </root>
